+++ iot-app/ViewControllers/ProfileViewController.swift
@@ -1,0 +1,46 @@
+//
+//  ProfileViewController.swift
+//  iot-app
+//
+//  Created by yokada on 2021/05/26.
+//
+
+import Amplify
+import AmplifyPlugins
+import UIKit
+
+class ProfileViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        // Do any additional setup after loading the view.
+        
+    }
+    
+    @IBAction func signOut(_ sender: Any) {
+        signOutLocally()
+    }
+    
+    func signOutLocally() {
+        Amplify.Auth.signOut() { result in
+            switch result {
+            case .success:
+                print("Successfully signed out")
+            case .failure(let error):
+                print("Sign out failed with error \(error)")
+            }
+        }
+    }
+    
+    /*
+    // MARK: - Navigation
+
+    // In a storyboard-based application, you will often want to do a little preparation before navigation
+    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+        // Get the new view controller using segue.destination.
+        // Pass the selected object to the new view controller.
+    }
+    */
+
+}
